@@ -1,7 +1,7 @@
 [package]
 name = "jymbol"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 authors = ["Jordan Danford <jordandanford@gmail.com>"]
 
 [dependencies]
@@ -9,5 +9,4 @@ anyhow = "1.0"
 chumsky = "0.8"
 gc = { version = "0.5", features = ["derive"] }
 im = "15"
-once_cell = "1.17"
 symbol_table = "0.3"
